+++ gems.rb
@@ -31,3 +31,4 @@ group :test do
 	gem "bake-test"
 	gem "bake-test-external"
 end
+